refact: rename :platform:animator's Animation package
