+++ platform/animator/src/androidTest/java/com/jeanbarrossilva/orca/platform/animator/animation/AnimatableTests.kt
@@ -13,13 +13,14 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package com.jeanbarrossilva.orca.platform.animator.animatable
+package com.jeanbarrossilva.orca.platform.animator.animation
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.remember
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
+import com.jeanbarrossilva.orca.platform.animator.animation.animatable.Animatable
 import org.junit.Rule
 import org.junit.Test
